requirements.txt
microsoft-bonsai-api
numpy
python-dotenv
